+++ emg.h
@@ -3,7 +3,7 @@
 
 // Standard header file, just look at emg.cpp for the implementation
 
-typedef void (*CallbackFunction)(double[], double, double, double[], double, double, double[], double, double);
+typedef void (*CallbackFunction)(float[], float, float, float[], float, float, float[], float, float);
 void setEmgSensor(CallbackFunction callback);
 void processFFT(CallbackFunction callback);
 void setupEMG();
